move validation to bokeh.core
