+++ bokeh/core/validation/check.py
