+++ bokeh/core/validation/decorators.py
