+++ bokeh/core/validation/exceptions.py
