+++ bokeh/core/validation/warnings.py
